Move OCR processors to different classes
