+++ lib/OcrProcessors/IOcrProcessor.php
@@ -34,6 +34,7 @@ interface IOcrProcessor {
 	 * @param File 				$file  				The file to be processed
 	 * @param WorkflowSettings 	$settings 			The settings to be used for this specific workflow
 	 * @param GlobalSettings 	$globalSettings 	The global settings configured for all OCR workflows on this system
+	 * @return OcrProcessorResult
 	 * @throws OcrNotPossibleException
 	 */
 	public function ocrFile(File $file, WorkflowSettings $settings, GlobalSettings $globalSettings) : OcrProcessorResult;
