+++ lib/OcrProcessors/PdfOcrProcessor.php
@@ -23,115 +23,9 @@ declare(strict_types=1);
 
 namespace OCA\WorkflowOcr\OcrProcessors;
 
-use Cocur\Chain\Chain;
-use OCA\WorkflowOcr\Exception\OcrNotPossibleException;
-use OCA\WorkflowOcr\Model\GlobalSettings;
-use OCA\WorkflowOcr\Model\WorkflowSettings;
 use OCA\WorkflowOcr\Wrapper\ICommand;
-use OCA\WorkflowOcr\Wrapper\IImageToPdfConverter;
-use OCP\Files\File;
 use Psr\Log\LoggerInterface;
 
-class PdfOcrProcessor implements IOcrProcessor {
-	/** @var array
-	 * Mapping for VUE frontend lang settings.
-	 * See also https://github.com/tesseract-ocr/tesseract/blob/main/doc/tesseract.1.asc#languages
-	 */
-	private static $langMapping = [
-		'de' => 'deu',
-		'en' => 'eng',
-		'fr' => 'fra',
-		'it' => 'ita',
-		'es' => 'spa',
-		'pt' => 'por',
-		'ru' => 'rus',
-		'chi' => 'chi_sim'
-	];
+class PdfOcrProcessor extends OcrMyPdfBasedProcessor {
 
-	/** @var ICommand */
-	private $command;
-
-	/** @var IImageToPdfConverter */
-	private $converter;
-
-	/** @var LoggerInterface */
-	private $logger;
-
-	public function __construct(ICommand $command, IImageToPdfConverter $converter, LoggerInterface $logger) {
-		$this->command = $command;
-		$this->converter = $converter;
-		$this->logger = $logger;
-	}
-
-	public function ocrFile(File $file, WorkflowSettings $settings, GlobalSettings $globalSettings): OcrProcessorResult {
-		if ($file->getMimeType() !== 'application/pdf') {
-			// Convert file to pdf. Here we assume that we're dealing with an image input
-			$pdfContent = $this->converter->convertToPdf($file->getContent());
-		} else {
-			$pdfContent = $file->getContent();
-		}
-		
-		$commandStr = 'ocrmypdf -q ' . $this->getCommandlineArgs($settings, $globalSettings) . ' - - | cat';
-
-		$this->command
-			->setCommand($commandStr)
-			->setStdIn($pdfContent);
-
-		$this->logger->debug('Running command: ' . $commandStr);
-
-		$success = $this->command->execute();
-		$errorOutput = $this->command->getError();
-		$stdErr = $this->command->getStdErr();
-		$exitCode = $this->command->getExitCode();
-
-		if (!$success) {
-			throw new OcrNotPossibleException('OCRmyPDF exited abnormally with exit-code ' . $exitCode . '. Message: ' . $errorOutput . ' ' . $stdErr);
-		}
-
-		if ($stdErr !== '' || $errorOutput !== '') {
-			// Log warning if ocrmypdf wrote a warning to the stderr
-			$this->logger->warning('OCRmyPDF succeeded with warning(s): {stdErr}, {errorOutput}', [
-				'stdErr' => $stdErr,
-				'errorOutput' => $errorOutput
-			]);
-		}
-
-		$ocrFileContent = $this->command->getOutput();
-
-		if (!$ocrFileContent) {
-			throw new OcrNotPossibleException('OCRmyPDF did not produce any output');
-		}
-
-		$this->logger->debug("OCR processing was successful");
-
-		return new OcrProcessorResult($ocrFileContent, "pdf");
-	}
-
-	private function getCommandlineArgs(WorkflowSettings $settings, GlobalSettings $globalSettings): string {
-		$args = [];
-			
-		// Language settings
-		if ($settings->getLanguages()) {
-			$langStr = Chain::create($settings->getLanguages())
-				->map(function ($langCode) {
-					return self::$langMapping[(string)$langCode] ?? null;
-				})
-				->filter(function ($l) {
-					return $l !== null;
-				})
-				->join('+');
-			$args[] = "-l $langStr";
-		}
-
-		// Remove background option (incompatible with redo-ocr)
-		$args[] = $settings->getRemoveBackground() ? '--remove-background' : '--redo-ocr';
-
-		// Number of CPU's to be used
-		$processorCount = intval($globalSettings->processorCount);
-		if ($processorCount > 0) {
-			$args[] = '-j ' . $processorCount;
-		}
-
-		return implode(' ', $args);
-	}
 }
