--- tests/Integration/ImageToPdfConverterTest.php
@@ -1,39 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * @copyright Copyright (c) 2022 Robin Windey <ro.windey@gmail.com>
- *
- *  @license GNU AGPL version 3 or any later version
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-namespace OCA\WorkflowOcr\Tests\Integration;
-
-use OCA\WorkflowOcr\Wrapper\ImageToPdfConverter;
-use Test\TestCase;
-
-class ImageToPdfConverterTest extends TestCase {
-	public function testConvertToPdf() {
-		$imagePath = realpath('./tests/Integration/data/testdocument.png');
-		$pngContent = file_get_contents($imagePath);
-		$imageToPdfConverter = new ImageToPdfConverter();
-		$result = $imageToPdfConverter->convertToPdf($pngContent);
-
-		$matchCnt = preg_match('/^%PDF-1.\d+/', $result);
-		$this->assertTrue($matchCnt === 1);
-	}
-}
